requirements.txt
aiosqlite==0.21.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
authlib==1.6.5
beartype==0.22.8
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.3.0
debugpy==1.8.17
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.123.10
fastmcp==2.13.3
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
graphviz==0.21
greenlet==3.2.4
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.1.0
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
kiwisolver==1.4.9
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.23.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
openai==2.7.1
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0

prompt-toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.2
ptyprocess==0.7.0
pure-eval==0.2.3
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pyee==13.0.0
pygments==2.19.2
pygraphviz==1.14
pyjwt==2.10.1
pymupdf==1.26.6
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlite-vec==0.1.6
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.1
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
